+++ isaacgymenvs/utils/rlgames_utils.py
@@ -27,10 +27,7 @@
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 from rl_games.common import env_configurations, vecenv
-from rl_games.common.algo_observer import AlgoObserver
-from rl_games.algos_torch import torch_ext
 from isaacgymenvs.utils.utils import set_seed
-import torch
 import numpy as np
 from typing import Callable
 
@@ -90,62 +87,6 @@ def get_rlgames_env_creator(
 
         return env
     return create_rlgpu_env
-
-
-class RLGPUAlgoObserver(AlgoObserver):
-    """Allows us to log stats from the env along with the algorithm running stats. """
-
-    def __init__(self):
-        pass
-
-    def after_init(self, algo):
-        self.algo = algo
-        self.mean_scores = torch_ext.AverageMeter(1, self.algo.games_to_track).to(self.algo.ppo_device)
-        self.ep_infos = []
-        self.direct_info = {}
-        self.writer = self.algo.writer
-
-    def process_infos(self, infos, done_indices):
-        assert isinstance(infos, dict), "RLGPUAlgoObserver expects dict info"
-        if isinstance(infos, dict):
-            if 'episode' in infos:
-                self.ep_infos.append(infos['episode'])
-
-            if len(infos) > 0 and isinstance(infos, dict):  # allow direct logging from env
-                self.direct_info = {}
-                for k, v in infos.items():
-                    # only log scalars
-                    if isinstance(v, float) or isinstance(v, int) or (isinstance(v, torch.Tensor) and len(v.shape) == 0):
-                        self.direct_info[k] = v
-
-    def after_clear_stats(self):
-        self.mean_scores.clear()
-
-    def after_print_stats(self, frame, epoch_num, total_time):
-        if self.ep_infos:
-            for key in self.ep_infos[0]:
-                    infotensor = torch.tensor([], device=self.algo.device)
-                    for ep_info in self.ep_infos:
-                        # handle scalar and zero dimensional tensor infos
-                        if not isinstance(ep_info[key], torch.Tensor):
-                            ep_info[key] = torch.Tensor([ep_info[key]])
-                        if len(ep_info[key].shape) == 0:
-                            ep_info[key] = ep_info[key].unsqueeze(0)
-                        infotensor = torch.cat((infotensor, ep_info[key].to(self.algo.device)))
-                    value = torch.mean(infotensor)
-                    self.writer.add_scalar('Episode/' + key, value, epoch_num)
-            self.ep_infos.clear()
-        
-        for k, v in self.direct_info.items():
-            self.writer.add_scalar(f'{k}/frame', v, frame)
-            self.writer.add_scalar(f'{k}/iter', v, epoch_num)
-            self.writer.add_scalar(f'{k}/time', v, total_time)
-
-        if self.mean_scores.current_size > 0:
-            mean_scores = self.mean_scores.get_mean()
-            self.writer.add_scalar('scores/mean', mean_scores, frame)
-            self.writer.add_scalar('scores/iter', mean_scores, epoch_num)
-            self.writer.add_scalar('scores/time', mean_scores, total_time)
 
 
 class RLGPUEnv(vecenv.IVecEnv):